requirements/base.txt
Babel==2.4.0
Django==1.11
django-annoying==0.10.3
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-datetime-widget==0.9.3
django-environ==0.4.3
django-model-utils==3.0.0
django-phonenumber-field==1.3.0
django-scheduler==0.8.5
django-user-accounts==2.0.1
django-versatileimagefield==1.7.1
djangorestframework==3.6.3
icalendar==3.11.4
jsonfield==2.0.2
olefile==0.44
phonenumberslite==8.5.2
Pillow==4.0.0
pinax-eventlog==1.1.2
pinax-theme-bootstrap==8.0.1
pinax-webanalytics==3.0.0
python-dateutil==2.6.0
pytz==2017.2
reportlab==3.4.0
six==1.10.0
